+++ src/client/src/pages/detail/form/stratigraphy/lithology/form/remarksFormSection.tsx
@@ -1,5 +1,5 @@
 import { FC } from "react";
-import { FormContainer, FormInput } from "../../../../../../../components/form/form.ts";
+import { FormContainer, FormInput } from "../../../../../../components/form/form.ts";
 
 interface RemarksFormSectionProps {
   fieldName?: string;